fixing testRedirector compile error by moving it into ExecutorService for now.
